requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.
pytest>=7.0,<9.0
python-dotenv>=1.0,<2.0
